Day 15 of learning (02)

+++ C/Chapter_04/problem_06.c
@@ -2,16 +2,16 @@
     int main(){
         int x = 0;
         int y;
-        int sum = 0;
+        float sum = 0;
         int n;
 
         for(int n=1; n<=5; n++){
             printf("Enter Temp of sensor %d: ", n);
             scanf("%d", &y);
-
             sum = sum + y;
+
         }
-        printf("The average temp of all sensors is %d.", sum/5);
+        printf("The average temp of all sensors is %f.", sum/5);
 
 
         return 0;
